requirements.txt
Django==3.0.8
django-heroku==0.3.1
gunicorn==20.0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0